moved special functionality to feature flags
