+++ scripts/feature-flags/theme-sheet.js
@@ -1,0 +1,89 @@
+import { getMetadata } from '../aem.js';
+
+/**
+ * Loads a theme spread sheet config.
+ * To use, create a design sheet with columns: Property, Value, Section, Block.
+ * add column 'design' to the metadata and set it to the path of the design sheet for your page.
+ */
+
+function addOverlayRule(ruleSet, selector, property, value) {
+  if (!ruleSet.has(selector)) {
+    ruleSet.set(selector, [`--${property}: ${value};`]);
+  } else {
+    ruleSet.get(selector).push(`--${property}: ${value};`);
+  }
+}
+
+export default async function loadThemeSpreadSheetConfig() {
+  const theme = getMetadata('design');
+  if (!theme) return;
+  // make sure the json files are added to paths.json first
+  const resp = await fetch(`/${theme}.json?offset=0&limit=500`);
+
+  if (resp.status === 200) {
+    // create style element that should be last in the head
+    document.head.insertAdjacentHTML('beforeend', '<style id="style-overrides"></style>');
+    const sheets = window.document.styleSheets;
+    const sheet = sheets.item(sheets.length - 1);
+    // load spreadsheet
+    const json = await resp.json();
+    const tokens = json.data || json.default.data;
+    // go through the entries and create the rule set
+    const ruleSet = new Map();
+    tokens.forEach((e) => {
+      const {
+        Property, Value, Section, Block,
+      } = e;
+      let selector = '';
+      if (Section.length === 0 && Block.length === 0) {
+        // :root { --<property>: <value>; }
+        addOverlayRule(ruleSet, ':root', Property, Value);
+      } else {
+        // define the section selector if set
+        if (Section.length > 0) {
+          selector = `main .section.${Section}`;
+        } else {
+          selector = 'main .section';
+        }
+        // define the block selector if set
+        if (Block.length) {
+          Block.split(',').forEach((entry) => {
+            // eslint-disable-next-line no-param-reassign
+            entry = entry.trim();
+            let blockSelector = selector;
+            // special cases: default wrapper, text, image, button, title
+            switch (entry) {
+              case 'default':
+                blockSelector += ' .default-content-wrapper';
+                break;
+              case 'image':
+                blockSelector += ` .default-content-wrapper img, ${selector} .block.columns img`;
+                break;
+              case 'text':
+                blockSelector += ` .default-content-wrapper p:not(:has(:is(a.button , picture))), ${selector} .columns.block p:not(:has(:is(a.button , picture)))`;
+                break;
+              case 'button':
+                blockSelector += ' .default-content-wrapper a.button';
+                break;
+              case 'title':
+                blockSelector += ` .default-content-wrapper :is(h1,h2,h3,h4,h5,h6), ${selector} .columns.block :is(h1,h2,h3,h4,h5,h6)`;
+                break;
+              default:
+                blockSelector += ` .block.${entry}`;
+            }
+            // main .section.<section-name> .block.<block-name> { --<property>: <value>; }
+            // or any of the spacial cases above
+            addOverlayRule(ruleSet, blockSelector, Property, Value);
+          });
+        } else {
+          // main .section.<section-name> { --<property>: <value>; }
+          addOverlayRule(ruleSet, selector, Property, Value);
+        }
+      }
+    });
+    // finally write the rule sets to the style element
+    ruleSet.forEach((rules, selector) => {
+      sheet.insertRule(`${selector} {${rules.join(';')}}`, sheet.cssRules.length);
+    });
+  }
+}
